requirements.txt
certifi==2019.3.9
Django==2.2
django-filebrowser-no-grappelli==3.7.8
django-tinymce4-lite==1.7.5
gunicorn==19.9.0
jsmin==2.2.2
Pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0